파일 정리 및 분리
