+++ module/wgangp.py
@@ -1,0 +1,3 @@
+import torch.nn as nn 
+import numpy as np
+
