+++ tools/image_processing.py
